v1.1

+++ src/engine/Mesh.hpp
@@ -1,12 +1,15 @@
 #pragma once
 #include <vector>
-#include <iostream>
 #include <cmath>
 #include <cassert>
 #include "math/Vec2.hpp"
 
+// Represents a simple 2D convex polygon mesh.
+// Points must be in counter-clockwise order.
+// Automatically recenters points around the centroid.
+// Computes outward-facing edge normals.
 struct Mesh {
-    std::vector<Vec2> points;       // Points defining the polygon in counter-clockwise order
+    std::vector<Vec2> points;
     std::vector<Vec2> normals;      
 
     Mesh(const std::vector<Vec2> pnts)
@@ -26,7 +29,7 @@ struct Mesh {
         normals.clear();
         int n = points.size();
 
-        for (size_t i = 0; i < n; ++i) {
+        for (int i = 0; i < n; ++i) {
             Vec2 p1 = points[i];
             Vec2 p2 = points[(i + 1) % n];
 
@@ -37,6 +40,9 @@ struct Mesh {
     }
 }; 
 
+
+// Global container for all meshes.
+// Used as a registry for quick access.
 namespace meshdata
 {
     inline std::vector<Mesh> meshes;
